Refactor functional test
- Refactor functional test logic into one place, `functional_impl.py`
- Tidy imports
